requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# MinIO S3-compatible storage
minio==7.2.3

# Image processing
Pillow==10.2.0
colorthief==0.2.1

# HTTP client
httpx==0.26.0

# Validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Background tasks
celery==5.3.6
redis==5.0.1

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6

# Web scraping
playwright==1.41.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# OpenAI API
openai==1.54.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
aiosqlite==0.19.0
